app/requirements.txt
asgiref
borb
bleach
autopep8
cachetools
certifi
chardet
charset-normalizer
coreapi
coreschema
coverage
Django
django-cors-headers
django-debug-toolbar
django-filter
django-multiselectfield
django-phonenumber-field
django-phonenumbers
djangorestframework
djangorestframework-simplejwt
drf-yasg
future
gunicorn
httplib2
idna
inflection
itypes
Jinja2
MarkupSafe
matplotlib
oauthlib
oauth2client
packaging
Pillow
# protobuf
psycopg2
pyasn1
pyasn1-modules
pycodestyle
Pygments
PyJWT
pyparsing
pyopenssl
redis
whitenoise
dj-database-url

